requirements.txt
asgiref==3.3.4
Django==3.2.1
django-bootstrap-form==3.4
django-filter==2.4.0
django-phonenumber-field==5.1.0
phonenumbers==8.12.22
Pillow==8.2.0
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
